S3Hook list_keys - Filter on each page rather on all keys
Currently the list_keys function first consolidates all the keys listed in a path before applying the filters.
In case of a large number of keys being present in the path and few keys ultimately being selected (such as due to the from / to datetime filter), this may lead to out of memory issues that could be avoided by bringing the filtering step at the time pages are iterated throughout.